Use Menhir's incremental API to recover from parse errors
Also provides a parser for strings, which is useful for testing and
future features.

The return type of parsing functions is now morally a (possibly empty)
list of parse errors and as much of the parse tree as could be recovered,
instead of either a successful parse or an error.

Added a test for parsing.