Don't add spurious whitespace in TeX-like rendering
We rename the badly named Render_verbatim to Render_TeX_like. More
importantly, the whitespace behaviour of Render_TeX_like.render is
changed as follows.

Previously, spaces would be inserted between every Forester node. This
results in incorrect rendering in some cases, including when we want to
emit \tag{...} to LaTeX. Now we do not add whitespace automatically between
nodes; however, when we encounter an uninterpeted control sequence
(which should be emitted as a TeX control sequence), we add spacing
depending on whether it is a *control word* or a *control symbol* ---
following the discussion in Chapter 3 of the TeXBook. According to
op.cit., spaces are ignored after control words and not ignored after
control symbols.

> When a space comes after a control word (an all-letter control
sequence), it is ignored by TeX. [...] But when a space comes after a
control symbol, it is truly a space.

I believe this behaviour is now as close to desirable as we can get.

This patch pertains to the discussion here:
https://lists.sr.ht/~jonsterling/forester-discuss/%3CB051B670-D3A9-4317-AFA1-675405EB171F@jonmsterling.com%3E

Closes: https://todo.sr.ht/~jonsterling/forester/49
Signed-off-by: Jon Sterling <jon@jonmsterling.com>