Vendor dream-html to remove dep. on 'dream' and thus 'ssl'
The 'ssl' package seems to be a consistent bad actor in regard to build
breakage wrt. the system openssl dependency (see
https://github.com/savonet/ocaml-ssl/issues/154). My new rule is that
the core Forester tool should not depend on this library until such time
as they adopt a practice that does not cause my build to irreparably
break following every patch release of the system openssl dependency.

I have proposed to change the dream-html library to not depend on dream
here: https://github.com/yawaramin/dream-html/issues/28. Will be very
happy to resume our dependency on dream-html if that idea is adopted.